// portfolio-manager-toolkit/src/interfaces/ITheme.ts
export interface ITheme {
  colors: {
    primary: string;
    primaryAlternate: string;
    secondary: string;
    secondaryAlternate: string;
    tertiary: string;
    tertiaryAlternate: string;
    background: string;
    darkText: string;
    lightText: string;
  };
  fontSizes: {
    sm: string;
    md: string;
    lg: string;
  };
}